CategoryService implementation added

+++ src/main/java/com/blog/domain/Category.java
@@ -16,7 +16,7 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "CATEGORY_ID")
-    private int categoryId;
+    private long categoryId;
 
     @Column(name = "CATEGORY_NAME")
     private String name;
